removed deprecated files
