+++ src/App.tsx
@@ -1,19 +1,15 @@
 import React, { useState } from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import Home from "./containers/Home";
-import Notes from "./containers/notes/Notes";
-import MediaGallery from "./containers/MediaGallery";
-import { TooltipProvider } from "@/components/ui/Tooltip";
-import { Toaster } from "@/components/ui/Toast";
+import GlobalStyles from "@/styles/GlobalStyles";
 
-import GlobalStyles from "./styles/GlobalStyles";
-import "./index.css";
-import NotesV2 from "@/features/notes/components/NotesV2";
-import navbarItems from "@/features/navbar/config/navbarItems";
-import Navbar from "@/features/navbar/components/Navbar";
-
+import "@/index.css";
 import { ThemeProvider } from "@/features/theme";
 import { Settings, SettingsProvider } from "@/features/settings";
+import { NotesV2 } from "@/features/notes";
+import { Toaster } from "@/shared/components/Toast";
+import { TooltipProvider } from "@/shared/components/Tooltip";
+import { Navbar, navbarItems } from "@/features/navbar";
+import { Feed } from "@/features/feed";
 
 const App: React.FC = () => {
   const [isLeftSidebarOpen, setIsLeftSidebarOpen] = useState(true);
@@ -37,20 +33,8 @@ const App: React.FC = () => {
               />
               <main className="flex-grow pt-10">
                 <Routes>
-                  <Route path="/" element={<Home />} />
-                  <Route
-                    path="/notes"
-                    element={
-                      <Notes
-                        isLeftSidebarOpen={isLeftSidebarOpen}
-                        isRightSidebarOpen={isRightSidebarOpen}
-                        toggleLeftSidebar={toggleLeftSidebar}
-                        toggleRightSidebar={toggleRightSidebar}
-                      />
-                    }
-                  />
-                  <Route path="/new-notes" element={<NotesV2 />} />
-                  <Route path="/media" element={<MediaGallery />} />
+                  <Route path="/" element={<Feed />} />
+                  <Route path="/notes" element={<NotesV2 />} />
                   <Route path="/settings" element={<Settings />} />
                 </Routes>
               </main>
