--- src/components/ThemeToggle.tsx
@@ -1,19 +1,0 @@
-import { Moon, Sun } from "lucide-react"
-import { Button } from "@/components/ui/Button"
-import { useTheme } from "@/components/ThemeProvider"
-
-export function ThemeToggle() {
-  const { setTheme, theme } = useTheme()
-
-  return (
-    <Button
-      variant="ghost"
-      size="icon"
-      onClick={() => setTheme(theme === "light" ? "dark" : "light")}
-    >
-      <Sun className="h-[1.2rem] w-[1.2rem] rotate-0 scale-100 transition-all dark:-rotate-90 dark:scale-0" />
-      <Moon className="absolute h-[1.2rem] w-[1.2rem] rotate-90 scale-0 transition-all dark:rotate-0 dark:scale-100" />
-      <span className="sr-only">Toggle theme</span>
-    </Button>
-  )
-}
