--- src/components/ui/Toast.tsx
@@ -1,5 +1,0 @@
-import { Toaster as SonnerToaster, toast as sonnerToast } from 'sonner';
-
-export const Toaster = SonnerToaster;
-
-export const toast = sonnerToast;
