--- src/containers/MediaGallery.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-
-const MediaGallery = () => (
-  <div className="container mx-auto p-4 mt-2">
-    <h2 className="text-2xl font-bold mb-4">Photos + Videos</h2>
-    <p>Check out your media gallery.</p>
-  </div>
-);
-
-export default MediaGallery;
