+++ src/features/feed/index.tsx
@@ -1,0 +1,5 @@
+import Feed from './Feed';
+
+export {
+  Feed
+};
