+++ src/features/navbar/config/navbarItems.ts
@@ -1,11 +1,9 @@
 import { NavbarItemProps } from "../components/NavbarItem";
-import { Home, Notebook, Image } from "lucide-react";
+import { Home, Notebook } from "lucide-react";
 
 const navbarItems: NavbarItemProps[] = [
   { to: "/", icon: Home, text: "Home" },
-  { to: "/notes", icon: Notebook, text: "Notes" },
-  { to: "/new-notes", icon: Notebook, text: "New Notes" },
-  { to: "/media", icon: Image, text: "Files" },
+  { to: "/notes", icon: Notebook, text: "Notes" }
 ];
 
 export default navbarItems;
