+++ src/features/navbar/index.tsx
@@ -1,0 +1,7 @@
+import Navbar from './components/Navbar';
+import navbarItems from './config/navbarItems';
+
+export {
+  Navbar,
+  navbarItems
+};
