+++ src/features/notes/index.tsx
@@ -1,0 +1,11 @@
+import NotesV2 from './components/NotesV2';
+import NoteEditor from './components/NoteEditor';
+import RelatedNotes from './components/RelatedNotes';
+import { useNotes } from './hooks/useNotes';
+
+export {
+  NotesV2,
+  NoteEditor,
+  RelatedNotes,
+  useNotes
+};
