+++ src/features/theme/components/ThemeToggle.tsx
@@ -1,6 +1,6 @@
 import { Moon, Sun } from "lucide-react"
-import { Button } from "@/components/ui/Button"
-import { useTheme } from "@/components/ThemeProvider"
+import { Button } from "@/shared/components/Button"
+import { useTheme } from "../hooks/useTheme"
 
 export function ThemeToggle() {
   const { setTheme, theme } = useTheme()
