+++ src/shared/components/ContextMenu/ContextMenu.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { cn } from '@/lib/utils';
+import { cn } from '@/shared/utils/utils';
 
 interface ContextMenuProps {
   x: number;
