+++ src/shared/components/Tabs/Tabs.tsx
@@ -1,0 +1,5 @@
+import * as TabsPrimitive from "@radix-ui/react-tabs";
+
+const Tabs = TabsPrimitive.Root;
+
+export default Tabs;
