+++ src/shared/components/Tabs/index.tsx
@@ -1,0 +1,5 @@
+export { default as Tabs } from './Tabs';
+export { default as TabsList } from './TabsList';
+export { default as TabsTrigger } from './TabsTrigger';
+export { default as TabsContent } from './TabsContent';
+export { default as CustomTabBar } from './CustomTabBar';
