+++ src/shared/components/Tooltip.tsx
@@ -1,7 +1,6 @@
 import * as React from "react";
 import * as TooltipPrimitive from "@radix-ui/react-tooltip";
-
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/utils/utils";
 
 const TooltipProvider = TooltipPrimitive.Provider;
