+++ src/shared/components/input.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/utils/utils";
 import { InputHTMLAttributes } from "react";
 
 const Input = React.forwardRef<
