+++ src/shared/hooks/useResizableSidebar.ts
