+++ src/shared/utils/utils.ts
@@ -1,0 +1,10 @@
+import { type ClassValue, clsx } from "clsx"
+import { twMerge } from "tailwind-merge"
+
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs))
+}
+
+export function getFolderName(path: string): string {
+  return path.split('/').pop() || path
+}
